Upgrade to Kibana3 c/o kibana-rack
Use a project called kibana-rack to provide us Kibana3 in the same way that
we ran Kibana2, protected by signon.

- https://github.com/tabolario/kibana-rack

This appears to Just Work. The one significant difference is that this app
is now also responsible for reverse proxying access Elasticsearch for a
handful of methods and endpoints:

- https://github.com/tabolario/kibana-rack/blob/v0.1.1/lib/kibana/rack/web.rb#L46-58

I don't think that's a big risk though, since it's:

- limited to people with production access
- a reasonbly limited number of methods/endpoints
- a separate ES cluster