Exclude router from list of erroring apps
Exclude the `router` application from the list of erroring apps. The
router acts as a reverse proxy for all requests to all other
applications, so it's not useful to show it in this dashboard.

Do this by changing the query to only use logs that are tagged with
`application`, which excludes logs from the router.

I also think this is more eloquent than using all logs not tagged
`nginx`.